12/4/17 update 1
I've created an animation for the Greenpeace volunteer. I've also figured out how to make the player stop when they reach the volunteer using the collider function in p5.play and then make the animation change when they collided with the volunteer. I continued exploring the p5.play references to get ideas of how to enable the player to pass the volunteer if certain parameters were met. I tried making the sprite disappear when conditions were met, but the player still collided with the empty space where the volunteer was. I then changed the coding so that when certain parameters were met, the volunteer became no longer something to collide with, but something to overlap.  The player now runs behind the volunteer, but I'm pretty sure that's just because the volunteer was drawn second. It's a cosmetic error I'm not too concerned with at this point.

Now I'm on to putting the backgrounds in an array and making each blocking character pop up only on certain backgrounds.