12/4/17 update 2
I have forgotten the semantics of arrays, but am trying through trial and error to use an array for the background. This one didn't work, so I'm going to create a test sketch and try to create a for loop in which the background changes each time the player's x value >width or <0.