12/4/17 update 4
The background works now and has been modified several times. I'm about to create a group for the sprites of the stoppers so they can be more correctly categorized and pop up on the right backgrounds. After that, it's deciding what sound level each sprite will be able to be passed with. 

If I've got the time, I want to make a timer and a fail page so the player won't win if they don't complete the map within a certain amount of time. I'll need to draw the success and fail pages.